requirements.txt
fastapi
uvicorn[standard]
requests
google-cloud-bigquery>=3.0.0 ; extra == "bigquery"
pytest
pytest-asyncio